Add files via upload
This Power BI file contains the assessment of week 2 which is being assigned in the internship program organized by EDUNET FOUNDATION in collaboration with AICTE & SHELL focusing on developing green technologies using AI. This was all about creating and understanding various statistical charts on Power BI for given datasets. These charts provide a visual representation of the data and help to access data for every state in India.